backend/requirements.txt
# annotated-types==0.7.0
# anyio==4.9.0
# certifi==2025.6.15
# charset-normalizer==3.4.2
# click==8.2.1
# colorama==0.4.6
# dnspython==2.7.0
# email_validator==2.2.0
# fastapi==0.115.14
# fastapi-cli==0.0.7
# filelock==3.18.0
# fsspec==2025.5.1
# h11==0.16.0
# httpcore==1.0.9
# httptools==0.6.4
# httpx==0.28.1
# huggingface-hub==0.33.2
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.6
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# mpmath==1.3.0
# networkx==3.5
# numpy==2.3.1
# orjson==3.10.18
# packaging==25.0
# pillow==11.3.0
# pydantic==2.11.7
# pydantic-extra-types==2.10.5
# pydantic-settings==2.10.1
# pydantic_core==2.33.2
# Pygments==2.19.2
# python-dotenv==1.1.1
# python-multipart==0.0.20
# PyYAML==6.0.2
# regex==2024.11.6
# requests==2.32.4
# rich==14.0.0
# rich-toolkit==0.14.8
# safetensors==0.5.3
# setuptools==80.9.0
# shellingham==1.5.4
# sniffio==1.3.1
# starlette==0.46.2
# sympy==1.14.0
# tokenizers==0.21.2
# torch==2.7.1
# tqdm==4.67.1
# transformers==4.53.0
# typer==0.16.0
# typing-inspection==0.4.1
# typing_extensions==4.14.0
# ujson==5.10.0
# urllib3==2.5.0
# uvicorn==0.35.0
# watchfiles==1.1.0
# websockets==15.0.1
# fastapi==0.110.2
# uvicorn[standard]==0.29.0
# python-multipart==0.0.9
# transformers==4.41.1
# torch==2.3.0
# Pillow==10.3.0


fastapi==0.115.14
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.35.0

transformers==4.41.1
torch==2.3.0 
Pillow
numpy==1.26.4